requirements.txt
streamlit
selenium
beautifulsoup4
google-generativeai
python-dotenv
